+++ tests/observable.test.js
@@ -1,8 +1,8 @@
 import { HTTP } from "../dist/aias";
 
-const mockup = true;
-HTTP.setMockup({ data: mockup });
-HTTP.setEventType("observable");
+// const mockup = true;
+// HTTP.setMockup({ data: mockup });
+// HTTP.setEventType("observable");
 
 const scientist = {
   firstname: "Galileo",
@@ -12,57 +12,57 @@ const scientist = {
 };
 
 test("get returns true", done => {
-  HTTP.get("http://httpbin.org/get", "text").subscribe(data => {
-    expect(data).toBe(true);
+  HTTP.observable.get("http://httpbin.org/get", "text").subscribe(data => {
+    expect(data).toContain("headers");
     done();
   });
 });
 
 test("post returns true", done => {
-  HTTP.post("http://httpbin.org/post", "text", scientist).subscribe(data => {
-    expect(data).toBe(true);
+  HTTP.observable.post("http://httpbin.org/post", "text", scientist).subscribe(data => {
+    expect(data).toContain("headers");
     done();
   });
 });
 
 test("put returns true", done => {
-  HTTP.put("http://httpbin.org/put", "text", scientist).subscribe(data => {
-    expect(data).toBe(true);
+  HTTP.observable.put("http://httpbin.org/put", "text", scientist).subscribe(data => {
+    expect(data).toContain("headers");
     done();
   });
 });
 
 test("delete returns true", done => {
-  HTTP.delete("http://httpbin.org/delete", "text").subscribe(data => {
-    expect(data).toBe(true);
+  HTTP.observable.delete("http://httpbin.org/delete", "text").subscribe(data => {
+    expect(data).toContain("headers");
     done();
   });
 });
 
-test("connect returns true", done => {
-  HTTP.connect("http://httpbin.org/connect", "text").subscribe(data => {
-    expect(data).toBe(true);
-    done();
-  });
-});
+// test("connect returns true", done => {
+//   HTTP.observable.connect("http://httpbin.org/connect", "text").subscribe(data => {
+//     expect(data).toContain("headers");
+//     done();
+//   });
+// });
 
-test("options returns true", done => {
-  HTTP.options("http://httpbin.org/options", "text").subscribe(data => {
-    expect(data).toBe(true);
-    done();
-  });
-});
+// test("options returns true", done => {
+//   HTTP.observable.options("http://httpbin.org/options", "text").subscribe(data => {
+//     expect(data).toContain("status");
+//     done();
+//   });
+// });
 
-test("trace returns true", done => {
-  HTTP.trace("http://httpbin.org/trace", "text").subscribe(data => {
-    expect(data).toBe(true);
-    done();
-  });
-});
+// test("trace returns true", done => {
+//   HTTP.observable.trace("http://httpbin.org/trace", "text").subscribe(data => {
+//     expect(data).toContain("headers");
+//     done();
+//   });
+// });
 
 test("patch returns true", done => {
-  HTTP.patch("http://httpbin.org/patch", "text", scientist).subscribe(data => {
-    expect(data).toBe(true);
+  HTTP.observable.patch("http://httpbin.org/patch", "text", scientist).subscribe(data => {
+    expect(data).toContain("headers");
     done();
   });
 });